built the layout in autolayout with constraints. works in iPhone, but doesn't look great on iPad.
Important Note: took a long time to figure out that i was having serious problems with autolayout due to "Safe Areas" being enabled within the project View Controller. make sure this is disabled for the rest of the project lessons, as they do not address the issue when dealing with Xcode 9 Beta-- it wasn't introduced as a feature until a later Beta build, and introduces some new complications when dealing with the newer iPhone 8, 8+, and iPhone X that isn't covered in these lessons which can really fuck up your layouts and make the lessons go off the rails. disable the option for now.